pyproject.toml
[project]
name = "customer-review-insights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "httpx>=0.25.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]
